Refactored components
- Handled responsiveness using vanilla flexbox
