+++ src/components/App/App.css
@@ -3,12 +3,6 @@
 	width: 100%;
 	height: 100%;
 }
-.navbar {
-	height: 70px;
-	display: flex;
-	justify-content: center;
-	align-items: center;
-}
 
 .body-container {
 	height: 100%;
@@ -19,47 +13,10 @@
 	align-items: center;
 }
 
-.page-bg {
-	background-color: #111111;
-	width: 100%;
-	height: 100%;
-	display: flex;
-	justify-content: center;
-}
-
 .main-container {
 	min-width: 850px;
 	padding: 24px;
 	display: flex;
 	justify-content: flex-start;
 	flex-direction: column;
-}
-
-.header-container {
-	width: 100%;
-	display: flex;
-	flex-direction: column;
-	justify-content: flex-start;
-	background-color: #000000;
-	align-items: center;
-	border-bottom: 1px solid #333333;
-}
-
-.header-inner-container {
-	min-width: 850px;
-	padding: 80px 24px 100px 24px;
-	display: flex;
-	justify-content: center;
-	flex-direction: column;
-	background-image: url(https://res.cloudinary.com/ds24tivvl/image/upload/v1619416018/Cloudworks/Group_22.png);
-	background-size: contain;
-	background-repeat: no-repeat;
-	background-position: right;
-}
-
-.action-container {
-	margin-top: -72px;
-	z-index: 99999;
-	display: flex;
-	justify-content: space-between;
 }
