+++ src/components/App/App.tsx
@@ -1,7 +1,9 @@
 import React, { PureComponent } from 'react';
 
-import { WorkloadListContainer } from '../WorkloadList';
-import { WorkloadFormContainer } from '../WorkloadForm';
+import Navbar from './Navbar';
+import Header from './Header';
+import WorkloadPage from '../Pages/Workload';
+
 import './App.css';
 
 
@@ -9,41 +11,10 @@ class App extends PureComponent {
   render() {
     return (
       <div className="app-root">
-        <div className="navbar">
-          <svg width="182" height="19" viewBox="0 0 182 19" fill="none" xmlns="http://www.w3.org/2000/svg">
-            <path d="M58.472 18.288C57.368 18.288 56.368 18.112 55.472 17.76C54.592 17.392 53.832 16.848 53.192 16.128C52.568 15.408 52.088 14.52 51.752 13.464C51.416 12.392 51.248 11.152 51.248 9.74399C51.248 8.33599 51.416 7.08799 51.752 5.99999C52.088 4.89599 52.568 3.97599 53.192 3.23999C53.832 2.48799 54.592 1.91999 55.472 1.53599C56.368 1.15199 57.368 0.95999 58.472 0.95999C59.976 0.95999 61.224 1.27999 62.216 1.91999C63.208 2.55999 64 3.52799 64.592 4.82399L61.88 6.26399C61.64 5.51199 61.248 4.91199 60.704 4.46399C60.16 3.99999 59.416 3.76799 58.472 3.76799C57.288 3.76799 56.344 4.16799 55.64 4.96799C54.952 5.76799 54.608 6.88799 54.608 8.32799V10.968C54.608 12.424 54.952 13.544 55.64 14.328C56.344 15.096 57.288 15.48 58.472 15.48C59.416 15.48 60.184 15.224 60.776 14.712C61.384 14.2 61.824 13.568 62.096 12.816L64.664 14.328C64.056 15.576 63.248 16.552 62.24 17.256C61.232 17.944 59.976 18.288 58.472 18.288Z" fill="white"/>
-            <path d="M70.2465 18C69.1905 18 68.4145 17.736 67.9185 17.208C67.4225 16.68 67.1745 15.936 67.1745 14.976V0.23999H70.2465V15.552H71.9025V18H70.2465Z" fill="white"/>
-            <path d="M79.1972 18.288C78.3012 18.288 77.4932 18.136 76.7732 17.832C76.0692 17.528 75.4612 17.088 74.9492 16.512C74.4532 15.936 74.0692 15.248 73.7972 14.448C73.5252 13.632 73.3892 12.72 73.3892 11.712C73.3892 10.704 73.5252 9.79999 73.7972 8.99999C74.0692 8.19999 74.4532 7.51999 74.9492 6.95999C75.4612 6.38399 76.0692 5.94399 76.7732 5.63999C77.4932 5.33599 78.3012 5.18399 79.1972 5.18399C80.0932 5.18399 80.9012 5.33599 81.6212 5.63999C82.3412 5.94399 82.9492 6.38399 83.4452 6.95999C83.9572 7.51999 84.3492 8.19999 84.6212 8.99999C84.8932 9.79999 85.0292 10.704 85.0292 11.712C85.0292 12.72 84.8932 13.632 84.6212 14.448C84.3492 15.248 83.9572 15.936 83.4452 16.512C82.9492 17.088 82.3412 17.528 81.6212 17.832C80.9012 18.136 80.0932 18.288 79.1972 18.288ZM79.1972 15.816C80.0132 15.816 80.6532 15.568 81.1172 15.072C81.5812 14.576 81.8132 13.848 81.8132 12.888V10.56C81.8132 9.61599 81.5812 8.89599 81.1172 8.39999C80.6532 7.90399 80.0132 7.65599 79.1972 7.65599C78.3972 7.65599 77.7652 7.90399 77.3012 8.39999C76.8372 8.89599 76.6052 9.61599 76.6052 10.56V12.888C76.6052 13.848 76.8372 14.576 77.3012 15.072C77.7652 15.568 78.3972 15.816 79.1972 15.816Z" fill="white"/>
-            <path d="M95.2406 15.912H95.1206C95.0086 16.232 94.8566 16.536 94.6646 16.824C94.4886 17.096 94.2566 17.344 93.9686 17.568C93.6966 17.792 93.3606 17.968 92.9606 18.096C92.5766 18.224 92.1286 18.288 91.6166 18.288C90.3206 18.288 89.3286 17.864 88.6406 17.016C87.9526 16.168 87.6086 14.96 87.6086 13.392V5.47199H90.6806V13.08C90.6806 13.944 90.8486 14.608 91.1846 15.072C91.5206 15.52 92.0566 15.744 92.7926 15.744C93.0966 15.744 93.3926 15.704 93.6806 15.624C93.9846 15.544 94.2486 15.424 94.4726 15.264C94.6966 15.088 94.8806 14.88 95.0246 14.64C95.1686 14.384 95.2406 14.088 95.2406 13.752V5.47199H98.3126V18H95.2406V15.912Z" fill="white"/>
-            <path d="M109.638 15.912H109.518C109.294 16.632 108.87 17.208 108.246 17.64C107.622 18.072 106.894 18.288 106.062 18.288C104.478 18.288 103.262 17.72 102.414 16.584C101.582 15.448 101.166 13.824 101.166 11.712C101.166 9.61599 101.582 8.00799 102.414 6.88799C103.262 5.75199 104.478 5.18399 106.062 5.18399C106.894 5.18399 107.622 5.39999 108.246 5.83199C108.87 6.24799 109.294 6.81599 109.518 7.53599H109.638V0.23999H112.71V18H109.638V15.912ZM107.046 15.744C107.782 15.744 108.398 15.568 108.894 15.216C109.39 14.864 109.638 14.36 109.638 13.704V9.71999C109.638 9.11199 109.39 8.62399 108.894 8.25599C108.398 7.88799 107.782 7.70399 107.046 7.70399C106.246 7.70399 105.598 7.97599 105.102 8.51999C104.622 9.04799 104.382 9.75999 104.382 10.656V12.816C104.382 13.712 104.622 14.424 105.102 14.952C105.598 15.48 106.246 15.744 107.046 15.744Z" fill="white"/>
-            <path d="M115.029 5.47199H117.981L119.253 10.632L120.237 14.928H120.333L121.485 10.632L122.949 5.47199H125.733L127.245 10.632L128.421 14.928H128.517L129.501 10.632L130.749 5.47199H133.605L130.173 18H126.861L125.253 12.48L124.317 9.07199H124.245L123.333 12.48L121.749 18H118.485L115.029 5.47199Z" fill="white"/>
-            <path d="M140.674 18.288C139.778 18.288 138.97 18.136 138.25 17.832C137.546 17.528 136.938 17.088 136.426 16.512C135.93 15.936 135.546 15.248 135.274 14.448C135.002 13.632 134.866 12.72 134.866 11.712C134.866 10.704 135.002 9.79999 135.274 8.99999C135.546 8.19999 135.93 7.51999 136.426 6.95999C136.938 6.38399 137.546 5.94399 138.25 5.63999C138.97 5.33599 139.778 5.18399 140.674 5.18399C141.57 5.18399 142.378 5.33599 143.098 5.63999C143.818 5.94399 144.426 6.38399 144.922 6.95999C145.434 7.51999 145.826 8.19999 146.098 8.99999C146.37 9.79999 146.506 10.704 146.506 11.712C146.506 12.72 146.37 13.632 146.098 14.448C145.826 15.248 145.434 15.936 144.922 16.512C144.426 17.088 143.818 17.528 143.098 17.832C142.378 18.136 141.57 18.288 140.674 18.288ZM140.674 15.816C141.49 15.816 142.13 15.568 142.594 15.072C143.058 14.576 143.29 13.848 143.29 12.888V10.56C143.29 9.61599 143.058 8.89599 142.594 8.39999C142.13 7.90399 141.49 7.65599 140.674 7.65599C139.874 7.65599 139.242 7.90399 138.778 8.39999C138.314 8.89599 138.082 9.61599 138.082 10.56V12.888C138.082 13.848 138.314 14.576 138.778 15.072C139.242 15.568 139.874 15.816 140.674 15.816Z" fill="white"/>
-            <path d="M149.229 18V5.47199H152.301V8.06399H152.421C152.501 7.72799 152.621 7.40799 152.781 7.10399C152.957 6.78399 153.181 6.50399 153.453 6.26399C153.725 6.02399 154.045 5.83199 154.413 5.68799C154.797 5.54399 155.237 5.47199 155.733 5.47199H156.405V8.37599H155.445C154.405 8.37599 153.621 8.52799 153.093 8.83199C152.565 9.13599 152.301 9.63199 152.301 10.32V18H149.229Z" fill="white"/>
-            <path d="M158.651 0.23999H161.723V10.944H161.843L163.499 8.68799L166.307 5.47199H169.787L165.467 10.32L170.243 18H166.595L163.379 12.384L161.723 14.184V18H158.651V0.23999Z" fill="white"/>
-            <path d="M175.997 18.288C174.797 18.288 173.789 18.088 172.973 17.688C172.157 17.272 171.437 16.704 170.813 15.984L172.685 14.16C173.149 14.688 173.653 15.104 174.197 15.408C174.757 15.712 175.397 15.864 176.117 15.864C176.853 15.864 177.381 15.736 177.701 15.48C178.037 15.224 178.205 14.872 178.205 14.424C178.205 14.056 178.085 13.768 177.845 13.56C177.621 13.336 177.229 13.184 176.669 13.104L175.421 12.936C174.061 12.76 173.021 12.376 172.301 11.784C171.597 11.176 171.245 10.296 171.245 9.14399C171.245 8.53599 171.357 7.99199 171.581 7.51199C171.805 7.01599 172.125 6.59999 172.541 6.26399C172.957 5.91199 173.453 5.64799 174.029 5.47199C174.621 5.27999 175.277 5.18399 175.997 5.18399C176.605 5.18399 177.141 5.23199 177.605 5.32799C178.085 5.40799 178.517 5.53599 178.901 5.71199C179.285 5.87199 179.637 6.07999 179.957 6.33599C180.277 6.57599 180.589 6.84799 180.893 7.15199L179.093 8.95199C178.725 8.56799 178.285 8.24799 177.773 7.99199C177.261 7.73599 176.701 7.60799 176.093 7.60799C175.421 7.60799 174.933 7.72799 174.629 7.96799C174.341 8.20799 174.197 8.51999 174.197 8.90399C174.197 9.31999 174.317 9.63999 174.557 9.86399C174.813 10.072 175.237 10.224 175.829 10.32L177.101 10.488C179.805 10.872 181.157 12.104 181.157 14.184C181.157 14.792 181.029 15.352 180.773 15.864C180.533 16.36 180.189 16.792 179.741 17.16C179.293 17.512 178.749 17.792 178.109 18C177.485 18.192 176.781 18.288 175.997 18.288Z" fill="white"/>
-            <path d="M18.2238 4.68381C18.2238 4.68381 9.8207 3.99858 8.63051 10.4594C7.96328 14.0813 10.6862 16.7688 13.2314 16.7688H34.0821C44.067 16.7688 45.1438 7.22898 35.5505 6.05429C35.5505 6.05429 27.0297 -3.6944 18.2238 4.68381Z" fill="white"/>
-            <path d="M0 8.12655C0 7.96912 0.127626 7.84149 0.285061 7.84149H17.5891C17.8382 7.84149 17.9674 8.13848 17.7977 8.32078L17.0016 9.17597C16.9477 9.23389 16.8721 9.2668 16.793 9.2668H0.285062C0.127627 9.2668 0 9.13917 0 8.98174V8.12655Z" fill="black"/>
-            <path d="M1.42529 10.9771C1.42529 10.8197 1.55292 10.6921 1.71035 10.6921H19.0144C19.2634 10.6921 19.3927 10.9891 19.223 11.1714L18.4269 12.0266C18.373 12.0845 18.2974 12.1174 18.2183 12.1174H1.71035C1.55292 12.1174 1.42529 11.9898 1.42529 11.8323V10.9771Z" fill="black"/>
-            <path d="M2.28052 14.1128C2.28052 13.9554 2.40814 13.8278 2.56558 13.8278H20.5831C20.8507 13.8278 20.9708 14.1633 20.7639 14.3332L20.0694 14.9033C20.0184 14.9451 19.9545 14.968 19.8886 14.968H2.56558C2.40814 14.968 2.28052 14.8404 2.28052 14.6829V14.1128Z" fill="black"/>
-            </svg>
-        </div>
+        <Navbar />
         <hr className="margin-0"/>
-        <div className="body-container">
-          <div className="header-container">
-            <div className="header-inner-container">
-              <h1 className="margin-0">Workload Manager</h1>
-              <p>Create, view and manage your workloads.</p>
-            </div>
-          </div>
-          <div className="page-bg">
-            <div className="main-container">
-              <div className="action-container">
-                  <WorkloadFormContainer />
-                  <WorkloadListContainer />
-              </div>
-            </div> 
-          </div>   
-        </div>
+        <Header />
+        <WorkloadPage />
       </div>
     );
   }
