+++ src/components/App/Header/Header.css
@@ -1,0 +1,35 @@
+.header-container {
+  padding: 0 40px;
+	background-color: #000000;
+  border-bottom: 1px solid #333333;
+  display: flex;
+  justify-content: center;
+}
+
+.header-inner-container {
+  width: 770px;
+	padding: 80px 0 100px 0;
+	background-image: url(https://res.cloudinary.com/ds24tivvl/image/upload/v1619416018/Cloudworks/Group_22.png);
+	background-size: contain;
+	background-repeat: no-repeat;
+	background-position: right;
+}
+
+/* Smartphones (portrait) ----------- */
+@media only screen and (max-width : 768px) {
+  .header-container {
+    padding: 0 20px;
+  }
+
+  .header-inner-container {
+    width: 100%;
+    flex-direction: column;
+  }
+}
+
+
+@media only screen and (min-width : 768px) and (max-width: 1024px) {
+  .header-inner-container {
+    width: 730px;
+  }
+}
