+++ src/components/App/Header/index.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import './Header.css';
+
+const Header = () => (
+  <div className="header-container">
+    <div className="header-inner-container">
+      <h1 className="margin-0">Workload Manager</h1>
+      <p>Create, view and manage your workloads.</p>
+    </div>
+  </div>
+)
+export default Header;
