+++ src/components/App/Navbar/Navbar.css
@@ -1,0 +1,6 @@
+.navbar {
+	height: 70px;
+	display: flex;
+	justify-content: center;
+	align-items: center;
+}
