+++ src/components/Pages/Workload/WorkloadForm/WorkloadForm.css
@@ -1,5 +1,4 @@
 .create-workload-container {
-  width: 370px;
   background: #000000;
   border: 1px solid #333333;
   box-sizing: border-box;
@@ -7,6 +6,7 @@
   display: flex;
   flex-direction: column;
   height: 600px;
+  margin-right: 16px;
 }
 
 .form-container {
@@ -49,7 +49,7 @@ input:focus-visible {
   height: 54px;
   width: 100%;
   font-size: 12px;
-  margin-top: 40px;
+  margin: 40px 0;
   font-weight: 600;
   text-transform: uppercase;
   transition: all 300ms ease-in-out;
@@ -63,4 +63,25 @@ button:hover {
   margin-top: 24px;
   display: flex;
   justify-content: space-between;
+}
+
+/* Smartphones (portrait) ----------- */
+@media only screen and (max-width : 768px) {
+  .create-workload-container {
+    margin-right: 0;
+    margin-bottom: 40px;
+    height: auto;
+  }
+}
+
+@media only screen and (min-width: 768px) and (max-width : 1024px) {
+  .create-workload-container {
+    width: 320px;
+  }
+}
+
+@media only screen and (min-width : 1024px) {
+  .create-workload-container {
+    width: 370px;
+  }
 }
