--- src/components/WorkloadForm/index.ts
@@ -1,2 +1,0 @@
-export { default } from './WorkloadForm';
-export * from './WorkloadForm';
