--- src/components/WorkloadItem/index.ts
@@ -1,2 +1,0 @@
-export { default } from './WorkloadItem';
-export * from './WorkloadItem';
