--- src/components/WorkloadList/index.ts
@@ -1,1 +1,0 @@
-export * from './WorkloadList';
